Move songbook reading inside patacrep wrapper
